backend/requirements.txt
@'
django
psycopg2-binary
python-dotenv
django-cors-headers
djangorestframework
'@ > requirements.txt
